+++ dds_gen/tests/structs.rs
@@ -58,7 +58,7 @@ fn module_generation() {
         r#"
         pub mod Game {
             pub mod Chess {
-                #[derive(Debug)]
+                #[derive(Debug, dust_dds::infrastructure::type_support::DdsType)]
                 pub enum ChessPiece {
                     Pawn,
                     Rook,
@@ -74,7 +74,7 @@ fn module_generation() {
                 }
             }
             pub mod Cards {
-                #[derive(Debug)]
+                #[derive(Debug, dust_dds::infrastructure::type_support::DdsType)]
                 pub enum Suit {
                     Spades,
                     Hearts,
@@ -111,7 +111,7 @@ fn nested_types() {
 
     let expected = syn::parse2::<File>(
         r#"
-            #[derive(Debug)]
+            #[derive(Debug, dust_dds::infrastructure::type_support::DdsType)]
             pub enum Presence {
                 Present,
                 NotPresent,
